refactor: update loaders to new collection struct
