+++ massdash/loaders/MzMLDataLoader.py
@@ -12,9 +12,7 @@ import pandas as pd
 from .access.MzMLDataAccess import MzMLDataAccess
 from .GenericSpectrumLoader import GenericSpectrumLoader
 # Structs
-from ..structs.TransitionGroup import TransitionGroup
-from ..structs.FeatureMap import FeatureMap
-from ..structs.TargetedDIAConfig import TargetedDIAConfig
+from ..structs import TransitionGroup, FeatureMap, TargetedDIAConfig, FeatureMapCollection, TopTransitionGroupFeatureCollection
 # Utils
 from ..util import LOGGER
 
@@ -111,7 +109,7 @@ class MzMLDataLoader(GenericSpectrumLoader):
         out_df = out_df.loc[:,~out_df.columns.duplicated()]
         return out_df
 
-    def loadFeatureMaps(self, pep_id: str, charge: int, config=TargetedDIAConfig) -> Dict[str, FeatureMap]:
+    def loadFeatureMaps(self, pep_id: str, charge: int, config=TargetedDIAConfig) -> FeatureMapCollection:
         '''
         Loads a dictionary of FeatureMaps (where the keys are the filenames) from the results file
 
@@ -121,13 +119,13 @@ class MzMLDataLoader(GenericSpectrumLoader):
         Returns:
             FeatureMap: FeatureMap object containing peak boundaries, intensity and confidence
         '''
-        out = {}
+        out = FeatureMapCollection()
         top_features = [ self.rsltsFile.getTopTransitionGroupFeature(basename(splitext(d.filename)[0]), pep_id, charge) for d in self.dataFiles]
         self.libraryFile.populateTransitionGroupFeatures(top_features)
         for d, t in zip(self.dataFiles, top_features):
             if t is None:
                 LOGGER.debug(f"No feature found for {pep_id} {charge} in {d.filename}")
-                return FeatureMap()
+                out[d.filename] =  FeatureMap()
             else:
                 out[d.filename] = d.reduce_spectra(t, config)
         return out
